requirements.txt
python-dotenv==0.21.1
python-telegram-bot==13.15
anthropic-cli==0.0.4